Rename `PipelineRegistry` to `ShaderCache`
Rename the `PipelineRegistry` object to `ShaderCache` to better reflect
the fact the cache only holds shaders, not pipelines. Also rename its
`configure()` method to `get_or_insert()` for clarity, as it does not do
any processing on the shader code.

Ensure `get_or_insert()` doesn't trigger change detection unnecessarily
on `Assets<Shader>` by passing the parameter as `&mut ResMut<T>` instead
of `&mut T`.
